% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfai.R
\docType{class}
\name{MFAI-class}
\alias{MFAI-class}
\title{Each MFAI object has a number of slots which store information. Key slots to access are listed below.}
\value{
MFAI class.
}
\description{
Each MFAI object has a number of slots which store information. Key slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{Y}}{numeric. The main data matrix.}

\item{\code{X}}{ANY. The auxiliary information data frame.}

\item{\code{Z}}{numeric. Estimated low-rank matrix, corresponding to posterior mean mu of Z in the MFAI model.}

\item{\code{W}}{numeric. Estimated low-rank matrix, corresponding to posterior mean nu of W in the MFAI model.}

\item{\code{tau}}{numeric. Precision parameter for each pair of factors.}

\item{\code{beta}}{numeric. Precision parameter for each Z factor.}

\item{\code{fX}}{numeric. Prior mean of Z, corresponding to f(X) part in the MFAI model.}

\item{\code{treeList}}{list. Fitted tree list, corresponding to f(.) function in the MFAI model.}

\item{\code{initialization}}{list. Initialization of the fitted model.}

\item{\code{boostingParameter}}{list. A list of options that control details of the rpart algorithm.}

\item{\code{.}}{}
}}

